Fix CI: repo-wide ruff lint drift, pin ruff version to prevent recurrence
CI's pip install pulled ruff 0.16.0 (unpinned dev dependency), which
enables several rules by default that weren't active in the ruff version
used for local checks before pushing 49c5f8c - none of the 29 flagged
issues were in the Phase 4 commit's own new files, they were pre-existing
repo-wide (mostly import-sort ordering and quoted-vs-bare type annotations
under `from __future__ import annotations`).

Applied `ruff --fix` for the 25 auto-fixable issues, then fixed the
remaining 4 by hand: Union[X, Y] -> X | Y in vtea_core.data.volume,
merged isinstance() calls in ParameterForm.get_values(), a redundant
int() around round() (verified round() on a numpy float64 already
returns a Python int, so the fix doesn't change behavior), and a
`# noqa: PYI034` on MeasurementStore.__enter__ instead of adopting
typing.Self, which needs Python 3.11+ while this package supports 3.10.

Pinned ruff>=0.16,<0.17 in both packages' dev extras so a future ruff
release can't silently fail CI on unrelated code the same way again.

+++ packages/vtea-napari/src/vtea_napari/widgets/gallery.py
@@ -25,7 +25,7 @@ _COLUMNS = 6
 class _ClickableThumbnail(QLabel):
     clicked = Signal()
 
-    def mousePressEvent(self, event):  # noqa: N802 - Qt override signature
+    def mousePressEvent(self, event):
         self.clicked.emit()
         super().mousePressEvent(event)
 
@@ -73,7 +73,7 @@ class GalleryWidget(QWidget):
         for position, object_id in enumerate(object_ids):
             if object_id not in indexed.index:
                 continue
-            row_center, col_center = (int(round(indexed.loc[object_id, c])) for c in spatial_columns)
+            row_center, col_center = (round(indexed.loc[object_id, c]) for c in spatial_columns)
             crop = _crop_2d(volume, row_center, col_center, crop_radius)
             if crop.size == 0:
                 continue
